Move logger to root
